Fix the code style

+++ contracts/ICommonsBudget.sol
@@ -3,5 +3,9 @@
 pragma solidity >=0.6.0 <0.9.0;
 
 interface ICommonsBudget {
-    function votePublished(bytes32 proposalID, uint validatorSize, uint64[] calldata voteCounts) external;
+    function votePublished(
+        bytes32 proposalID,
+        uint256 validatorSize,
+        uint64[] calldata voteCounts
+    ) external;
 }
